backport refactorings from #1284
